:recycle: Extract LO CUD to module
Create, Update, and Delete for learning objects was moved to it's own module. The module is not yet fully standalone, as it still uses the same datastore as the rest of the app and is depended on by a route in the AuthRouteDriver.
